item texts and show items mode

+++ handlers/handlers_item_edit_inline_buttons.py
@@ -12,12 +12,11 @@ from handlers.handlers_folder import show_folders
 from handlers.handlers_item import show_item
 from handlers.handlers_item_inline_buttons import get_item_body_and_current_markup
 from load_all import bot, dp
-from models.item_model import Item
+from models.item_model import Item, INVISIBLE_CHAR
 from utils.data_manager import get_data, set_data
 from utils.utils_ import invisible_char
 from utils.utils_button_manager import item_edit_buttons, create_general_reply_markup, \
-    get_edit_item_title_keyboard, cancel_edit_item_button, get_edit_item_text_keyboard, delete_page_inline_button, \
-    get_text_pages_buttons
+    get_edit_item_title_keyboard, get_edit_item_text_keyboard, get_text_pages_buttons, get_edit_page_buttons
 from utils.utils_items_db import util_edit_item
 from utils.utils_items_reader import get_item
 from utils.utils_parse_mode_converter import to_markdown_text, preformat_text, full_escape_markdown, \
@@ -105,7 +104,7 @@ async def edit_item_text_handler(call: CallbackQuery, state: FSMContext):
     inline_markup, page = await get_inline_markup_and_page(call.message)
 
     edit_item_messages = []
-    if item.get_text():
+    if item.pages_count() and (item.pages_count() > 1 or item.page_not_empty(0)):
         edit_item_messages.append(
             await bot.send_message(chat_id=user_id,
                                    text=get_instruction_copy_edit_text(item, page),
@@ -116,7 +115,7 @@ async def edit_item_text_handler(call: CallbackQuery, state: FSMContext):
 
     await asyncio.sleep(0.3)
 
-    buttons = get_edit_item_text_keyboard(item.text)
+    buttons = get_edit_item_text_keyboard(item)
     markup = create_general_reply_markup(buttons)
 
     edit_item_messages.append(
@@ -125,8 +124,10 @@ async def edit_item_text_handler(call: CallbackQuery, state: FSMContext):
                                reply_markup=markup)
     )
 
-    await bot.delete_message(chat_id=user_id, message_id=call.message.message_id)
-    data.pop('bot_message')
+    if item.pages_count() and (item.pages_count() > 1 or item.page_not_empty(0)):
+        await bot.delete_message(chat_id=user_id, message_id=call.message.message_id)
+        data.pop('bot_message')
+
     data['edit_item_messages'] = edit_item_messages
     await set_data(user_id, data)
     await state.update_data(item_text_page=page)
@@ -148,9 +149,32 @@ async def get_inline_markup_and_page(message: Message):
             callback_data.action = f'{action}_edit'
             btn.callback_data = callback_data.pack()
 
-        buttons = [current_markup.inline_keyboard[0], [delete_page_inline_button]]
+        buttons = [current_markup.inline_keyboard[0], get_edit_page_buttons()]
         inline_markup = InlineKeyboardMarkup(row_width=3, inline_keyboard=buttons)
     return inline_markup, page
+
+
+@router.callback_query(F.data == "insert_page")
+async def insert_text_page_handler(call: CallbackQuery, state: FSMContext):
+    user_id = call.from_user.id
+    data = await get_data(user_id)
+    item_id = data.get('item_id')
+
+    item: Item = await get_item(user_id, item_id)
+    inline_markup, page = await get_inline_markup_and_page(call.message)
+
+    item.insert_page(page_index=page)
+    result = await util_edit_item(user_id, item_id, item)
+
+    buttons = [get_text_pages_buttons(user_id, item, page, mode='_edit'), get_edit_page_buttons()]
+    inline_markup = InlineKeyboardMarkup(row_width=3, inline_keyboard=buttons)
+
+    await call.message.edit_text(
+        text=get_instruction_copy_edit_text(item, page),
+        reply_markup=inline_markup,
+        parse_mode=ParseMode.MARKDOWN_V2
+    )
+    await call.answer()
 
 
 @router.callback_query(F.data == "remove_page")
@@ -201,13 +225,13 @@ async def action_remove_text_page_handler(call: CallbackQuery, state: FSMContext
     if action == 'yes':
         item.remove_page(page)
         result = await util_edit_item(user_id, item_id, item)
-        if result:
+        if result and page > 0:
             page -= 1
             await state.update_data(item_text_page=page)
 
     inline_markup = None
     if item.pages_count() > 1:
-        buttons = [get_text_pages_buttons(user_id, item, page, mode='_edit'), [delete_page_inline_button]]
+        buttons = [get_text_pages_buttons(user_id, item, page, mode='_edit'), get_edit_page_buttons()]
         inline_markup = InlineKeyboardMarkup(row_width=3, inline_keyboard=buttons)
 
     await call.message.edit_text(
@@ -221,7 +245,7 @@ async def action_remove_text_page_handler(call: CallbackQuery, state: FSMContext
 def get_instruction_copy_edit_text(item: Item, page_number: int):
     entity = f'страницы {(page_number + 1)}' if len(item.text) > 1 else 'записи'
     return f"_Нажмите на текущий текст {entity}, чтобы скопировать:_ ↙️" \
-           f"\n\n`{markdown_without_code(item.get_text(page_number))}`"
+           f"\n\n`{markdown_without_code(item.get_text(page_number))}\n{INVISIBLE_CHAR}`"
 
 
 def get_instruction_new_edit_text(item: Item):
